Add files via upload
I've had this project in the works as an attempt to create my perfect Windows calculator. I don't like Windows calculator that simulate a keyboard (what's the point if you don't get the tactile feedback), so in Windows I just prefer the ability to type and evaluate expressions.
Unfortunately, although I was unsatisfied with other products I tried, I was never able to properly define my ideal features wish list and/or how to implement them in an elegant and user friendly way. I did implement most (if not all) of my original ideas, but I quickly ran out of ideas for new features or improvements.
At work, I mostly do calculations in Excel (no comments!), and at home, I mostly use my HP48SX as I like to punch on the keyboard.
As a result, this project has been in limbo for almost 2 years now, and I have no plan to update it, hence I'm just releasing it as it.
If someone can help contribute not just in code but also in ideas, feel free to do so. Of course, you can fork your own branch, and I'm open to merging contributions.